(GH-269) Add acceptance test for Workspace Symbols
Previously the Workspace Symbols request was not tested as part of the
acceptance suite.  This commit adds a Workspace symbols request test and
refactors the mock editor client to make it easier to be consumed during
development.
