+++ spec/languageserver/fixtures/control_repos/valid/site/profile/manifests/editorservices.pp
@@ -1,0 +1,3 @@
+class profile::editorservices {
+  # Empty class
+}
